prevent file rename when new nick is the same as old nick.

+++ src/groupchat_control.py
@@ -1006,7 +1006,7 @@ class GroupchatControl(ChatControlBase):
 						files[os.path.join(path, puny_nick + ext)] = \
 							os.path.join(path, puny_new_nick + ext)
 					for old_file in files:
-						if os.path.exists(old_file):
+						if os.path.exists(old_file) and old_file != files[old_file]:
 							if os.path.exists(files[old_file]):
 								# Windows require this
 								os.remove(files[old_file])
